Moved the StartingGun into the vertx module where it is used.
