+++ api/src/main/java/module-info.java
@@ -39,7 +39,6 @@ module com.microsoft.gctoolkit.api {
     exports com.microsoft.gctoolkit.io;
     exports com.microsoft.gctoolkit.jvm;
     exports com.microsoft.gctoolkit.time;
-    exports com.microsoft.gctoolkit.util.concurrent;
 
     uses com.microsoft.gctoolkit.aggregator.Aggregation;
     uses com.microsoft.gctoolkit.jvm.JavaVirtualMachine;
