+++ vertx/src/main/java/com/microsoft/gctoolkit/vertx/internal/util/concurrent/StartingGun.java
@@ -1,4 +1,4 @@
-package com.microsoft.gctoolkit.util.concurrent;
+package com.microsoft.gctoolkit.vertx.internal.util.concurrent;
 
 import java.util.concurrent.locks.AbstractQueuedSynchronizer;
